+++ src/sphinx_codeautolink/__init__.py
@@ -4,6 +4,7 @@ from pathlib import Path as _Path
 
 from sphinx.application import Sphinx
 from .extension import backref, directive, SphinxCodeAutoLink
+from .extension.block import clean_pycon  # NOQA
 
 _version_file = _Path(_os.path.realpath(__file__)).parent / "VERSION"
 __version__ = _version_file.read_text().strip()