+++ tests/__init__.py
@@ -4,3 +4,6 @@ import sphinx_codeautolink
 class TestPackage:
     def test_version(self):
         assert sphinx_codeautolink.__version__
+
+    def test_clean_pycon_public(self):
+        assert sphinx_codeautolink.clean_pycon